小数点位置で座標指定出来るようにする。

+++ graphics/graphics.h
@@ -44,7 +44,7 @@ void DrawLinearGradient(int16_t x1, int16_t y1, pixel_t v1,
 						bool dithering);
 
 void DrawRadialGradient(
-	int16_t cx, int16_t cy, uint16_t diameter,
+	float cx, float cy, float diameter,
 	const ClippingRect& clippingRect,
 	const uint16_t* distanceTable, uint8_t distanceTableShifts,
 	const pixel_t* pixelTable, uint8_t pixelTableShits,
